+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Change Log
 
 - added SOCK_NODELAY config item to be able to turn the TCP_NODELAY socket option on (default is off).
 - little cleanup of the benchmark example
+- added timezones example
 
 
 **Pyro 4.35**